+++ UnitClasses/HealUnit.h
@@ -7,7 +7,7 @@
 class HealUnit: public Unit
 {
 public:
-	HealUnit(Game*, int, int, int);
+	HealUnit(Game*, double, int, int);
 	void attack();
 };
